added basecode for welcome modal
